pyproject.toml
[tool.poetry]
name = "verteilte-backend-dhbw"
version = "0.1.0"
description = ""
authors = ["TimoFeucht <timo@feucht-web.de>", "Mick Eisebraun"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.28"
pydantic = "^2.6.3"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
psycopg2 = "^2.9.9"
requests = "^2.31.0"
sqlalchemy-rqlite = "^1.2"
pymongo = "^4.6.3"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
